some more package renaming
