+++ modules/client-scala/src/test/scala/be/wegenenverkeer/atomium/client/FeedIteratorFixture.scala
@@ -1,9 +1,8 @@
 package be.wegenenverkeer.atomium.client
 
 import be.wegenenverkeer.atomium.client.FeedEntryIterator.Implicits._
-import be.wegenenverkeer.atomium.client.async.AsyncFeedEntryIterator
 import be.wegenenverkeer.atomium.format.{Feed, Link, Url}
-import be.wegenenverkeer.atomium.server.{FeedService, MemoryFeedStore, Context}
+import be.wegenenverkeer.atomium.server.{Context, FeedService, MemoryFeedStore}
 import org.joda.time.DateTime
 
 import scala.util.{Success, Try}
